Fix `CircuitMerger` with unowned bits (#328)
* fix CI

* Fix `CircuitMerger` with unowned bits

The `Bit.index` attribute has been deprecated since Qiskit Terra 0.17
(April 2021), and was unreliable.  A set of bugfixes for other parts of
QPY in Terra 0.23 caused this information to be erased.  Even if Terra
rolls back the change temporarily, it will be removed in a future
release, as `Bit.index` is due to be removed entirely.

Co-authored-by: Takashi Imamichi <imamichi@jp.ibm.com>
Co-authored-by: Kyle Ross <Kyle.D.Ross@ibm.com>