Suppress deprecation warning from IBMQ QPY imports (#330)
* Temporarily suppress deprecation warnings from IBMQ

`qiskit-ibmq-provider` is still using deprecated paths to import QPY,
which causes test failures.  Since `ibmq-provider` is itself deprecated,
we are unkeen to push further changes to it, and intend to fully
replace it.  This just suppresses the warnings so the NTC programs can
be redeployed to working versions sooner.

* Force Travis to use newer Python 3.9

Travis' cached Python versions were causing the '3.9' specifier to mean
'3.9.0', which is very old.  Numba was having problems importing under
these conditions, which is required for
`qiskit-machine-learning[sparse]` and optionally used by `qiskit-nature`
if `sparse` is available.

This commit may not fix the problem; it's just exploratory, because I
can't reproduce failures to import `numba` locally.