Use marginal_distribution() instead of marginal_counts() (#349)
This commit updates the qiskit marginalization function used in the qasm3 runner
program. Qiskit has 2 different function available marginal_counts() [1] which is designed
to work with Counts or Result objects and does not respect ordering of the indices passed
in. The other is marginal_distribution() [2] which is designed to work with Counts,
Probabilities, QuasiProbabilities, or a dictionary with string keys and integer or
float values that respects the order of the indices passed in (so it can be used to
permute the bits). In general if you're not working with a Results object it is better
to use marginal_distribution() because it is written primarily in rust and 10-100x faster
than marginal_counts().

[1] https://qiskit.org/documentation/stubs/qiskit.result.marginal_counts.html
[2] https://qiskit.org/documentation/stubs/qiskit.result.marginal_distribution.html

Co-authored-by: Mariana Bernagozzi <Mariana.Bernagozzi@ibm.com>